folder rename: libs->libtools
