+++ src/wxtilesdeckgl.ts
@@ -1,8 +1,8 @@
 import './wxtilesdeckgl.css';
 
-export { setupWxTilesLib, setWxTilesLogging, createWxTilesLayerProps, CreateProps, WxServerVarsStyleType } from './libs/libTools';
+export { setupWxTilesLib, setWxTilesLogging, createWxTilesLayerProps, CreateProps, WxServerVarsStyleType } from './libtools/libTools';
 
-export { WxTilesLayerManager, createDeckGlLayer } from './libs/createDeckGlLayer';
+export { WxTilesLayerManager, createDeckGlLayer } from './libtools/createDeckGlLayer';
 
 export { WxTilesLayer } from './layers/WxTilesLayer';
 export { DebugTilesLayer } from './layers/DebugTilesLayer';
